Allow argument inference for non-quote types.
It is not yet possible to determine if a particular variable will end up
bound to a function or a plain value ahead of time, so it defaults to
treating everything as a value, rather than applying functions. This can
be fixed by adding a generic "any-to-any" function type annotation.